Move settings_dialog.py into RimSort.views
