+++ RimSort/views/settings_dialog.py
